backend/pyproject.toml
[project]
name = "catasto-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-gis>=1.2.1",
    "pillow>=12.2.0",
    "psycopg[binary]>=3.3.4",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
